+++ src/targets/gpu/device/argmin.cpp
@@ -12,7 +12,7 @@ inline namespace MIGRAPHX_INLINE_NS {
 namespace gpu {
 namespace device {
 
-void argmin(hipStream_t stream, const argument& result, const argument& arg, int axis)
+void argmin(hipStream_t stream, const argument& result, const argument& arg, int64_t axis)
 {
     arg_op(argmin_op{}, stream, result, arg, axis);
 }
